+++ src/services/statefuls/cache/manager.rs
@@ -1,15 +1,16 @@
-use super::cache_actor::{CacheActor, CacheCommand, CacheCommandSender};
-use super::ttl_manager::{TtlActor, TtlSchedulerInbox};
+use super::actor::{CacheActor, CacheCommand, CacheCommandSender};
+use crate::services::statefuls::cache::ttl::manager::TtlSchedulerInbox;
 use super::CacheEntry;
 use crate::services::stream_manager::query_io::QueryIO;
 
-use crate::services::statefuls::persist::persist_actor::SaveCommand;
+use crate::services::statefuls::persist::actor::SaveCommand;
 use crate::services::statefuls::persist::DumpFile;
 use anyhow::Result;
 use std::time::SystemTime;
 use std::{hash::Hasher, iter::Zip};
 use tokio::sync::mpsc;
 use tokio::sync::oneshot::Sender;
+use crate::services::statefuls::cache::ttl::actor::TtlActor;
 
 type OneShotSender<T> = tokio::sync::oneshot::Sender<T>;
 type OneShotReceiverJoinHandle<T> =
