+++ src/services/statefuls/persist/endec/encoder/encoding_processor.rs
@@ -1,7 +1,7 @@
 use crate::services::statefuls::cache::CacheEntry;
 use crate::services::stream_manager::interface::TWrite;
 
-use crate::services::statefuls::persist::persist_actor::SaveCommand;
+use crate::services::statefuls::persist::actor::SaveCommand;
 use anyhow::Result;
 
 use std::collections::VecDeque;
