+++ src/services/stream_manager/request_controller/client/mod.rs
@@ -1,9 +1,9 @@
 use crate::services::config::manager::ConfigManager;
 use crate::services::config::ConfigResponse;
-use crate::services::statefuls::cache::cache_manager::CacheManager;
-use crate::services::statefuls::cache::ttl_manager::TtlSchedulerInbox;
+use crate::services::statefuls::cache::manager::CacheManager;
+use crate::services::statefuls::cache::ttl::manager::TtlSchedulerInbox;
 use crate::services::statefuls::persist::endec::encoder::encoding_processor::SavingProcessor;
-use crate::services::statefuls::persist::persist_actor::PersistActor;
+use crate::services::statefuls::persist::actor::PersistActor;
 use crate::services::stream_manager::interface::TCancellationWatcher;
 use crate::services::stream_manager::query_io::QueryIO;
 use arguments::ClientRequestArguments;
